Rename proto package to avoid protobuf collision
Rename the internal proto package to protocol to avoid protobuf
collision with the client proto package.
