+++ server/metadata.go
@@ -17,7 +17,7 @@ import (
 
 	client "github.com/liftbridge-io/liftbridge-api/go"
 
-	"github.com/liftbridge-io/liftbridge/server/proto"
+	proto "github.com/liftbridge-io/liftbridge/server/protocol"
 )
 
 const (
