+++ server/protocol/internal.proto
@@ -1,5 +1,5 @@
 syntax = "proto3";
-package proto;
+package protocol;
 
     
 message ServerState {
