+++ server/replicator.go
@@ -10,7 +10,7 @@ import (
 	"github.com/nats-io/nats.go"
 
 	"github.com/liftbridge-io/liftbridge/server/commitlog"
-	"github.com/liftbridge-io/liftbridge/server/proto"
+	proto "github.com/liftbridge-io/liftbridge/server/protocol"
 )
 
 const (
